+++ aplicacoes/gpa-mae/src/main/java/br/ufc/quixada/npi/gpa/model/QuestionarioAuxilioMoradia.java
@@ -39,7 +39,7 @@ public class QuestionarioAuxilioMoradia {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long id;
+	private Integer id;
 
 	@ManyToOne
 	private SelecaoBolsa selecaoBolsa;
@@ -173,11 +173,11 @@ public class QuestionarioAuxilioMoradia {
 	@NotEmpty(message = "Campo obrigatório")
 	private String justificativa;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
